Moved all patches to their own directory
