+++ src/.patches/worker/shared.ts
