+++ src/boot_worker_app.ts
@@ -1,6 +1,6 @@
 import { platform, provide, ApplicationRef, ComponentRef, Injector } from "angular2/core";
 import { WORKER_APP_PLATFORM, WORKER_APP_APPLICATION } from "angular2/platform/worker_app";
-import { WORKER_APP_ROUTER, initRouter } from './.worker/worker';
+import { WORKER_APP_ROUTER, initRouter } from './.patches/worker/worker';
 import { APP_BASE_HREF, Router } from 'angular2/router';
 import { App } from './app/app';
